LIBDRUM-701. Moved community group implementation to stock.
https://issues.umd.edu/browse/LIBDRUM-701
