+++ src/app/core/shared/community-group.model.spec.ts
