+++ src/app/core/shared/community-group.model.ts
@@ -2,13 +2,13 @@ import { autoserialize, deserialize, inheritSerialization } from 'cerialize';
 import { Observable } from 'rxjs';
 import { ResourceType } from 'src/app/core/shared/resource-type';
 import { excludeFromEquals } from 'src/app/core/utilities/equals.decorators';
-import { link, typedObject } from '../../../../../app/core/cache/builders/build-decorators';
-import { CacheableObject } from '../../../../../app/core/cache/cacheable-object.model';
-import { PaginatedList } from '../../../../../app/core/data/paginated-list.model';
-import { RemoteData } from '../../../../../app/core/data/remote-data';
-import { Community } from '../../../../../app/core/shared/community.model';
-import { COMMUNITY } from '../../../../../app/core/shared/community.resource-type';
-import { HALLink } from '../../../../../app/core/shared/hal-link.model';
+import { link, typedObject } from '../cache/builders/build-decorators';
+import { CacheableObject } from '../cache/cacheable-object.model';
+import { PaginatedList } from '../data/paginated-list.model';
+import { RemoteData } from '../data/remote-data';
+import { Community } from './community.model';
+import { COMMUNITY } from './community.resource-type';
+import { HALLink } from './hal-link.model';
 import { COMMUNITY_GROUP } from './community-group.resource-type';
 
 @typedObject
