+++ src/themes/drum/eager-theme.module.ts
@@ -13,7 +13,7 @@ import { UntypedItemComponent } from './app/item-page/simple/item-types/untyped-
 import { UmdHeaderComponent } from './app/umd-header/umd-header.component';
 import { FileSectionComponent } from './app/item-page/simple/field-components/file-section/file-section.component';
 import { BitstreamDownloadCounterComponent } from './app/bitstream-download-counter/bitstream-download-counter.component';
-import { CommunityGroup } from './app/core/shared/community-group.model';
+import { CommunityGroup } from '../../app/core/shared/community-group.model';
 import { HomePageComponent } from './app/home-page/home-page.component';
 import { StatisticsModule } from 'src/app/statistics/statistics.module';
 import { CommunityListPageModule } from 'src/app/community-list-page/community-list-page.module';
